requirements.txt
dateparser==1.0.0
emoji==1.2.0
geopy==2.1.0
google==3.0.0
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
gspread==5.1.1
nextcord>=2.0.0,<3
oauth2client==4.1.3
pandas==1.4.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==1.4.25
